:construction: Rename functions and files & move functions!
Rename files and function for better comprehension, and shift functions to new files!
